+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleDataprocService.scala
@@ -4,6 +4,8 @@ package mock
 import cats.syntax.all._
 import cats.effect.IO
 import cats.mtl.Ask
+import com.google.api.gax.longrunning.OperationSnapshot
+import com.google.api.gax.rpc.StatusCode
 import com.google.cloud.compute.v1.Operation
 import com.google.cloud.dataproc.v1.{Cluster, ClusterOperationMetadata}
 import org.broadinstitute.dsde.workbench.model.TraceId
@@ -15,9 +17,16 @@ class BaseFakeGoogleDataprocService extends GoogleDataprocService[IO] {
     region: RegionName,
     clusterName: DataprocClusterName,
     createClusterConfig: Option[CreateClusterConfig]
-  )(implicit ev: Ask[IO, TraceId]): IO[ClusterOperationMetadata] =
+  )(implicit ev: Ask[IO, TraceId]): IO[OperationSnapshot] =
     IO.pure(
-      ClusterOperationMetadata.newBuilder().setClusterName(clusterName.value).setClusterUuid("clusterUuid").build()
+      new OperationSnapshot {
+        override def getName: String = "opName"
+        override def getMetadata: AnyRef = ???
+        override def isDone: Boolean = true
+        override def getResponse: AnyRef = ???
+        override def getErrorCode: StatusCode = null
+        override def getErrorMessage: String = null
+      }
     )
 
   override def stopCluster(project: GoogleProject,
@@ -44,11 +53,11 @@ class BaseFakeGoogleDataprocService extends GoogleDataprocService[IO] {
                              numPreemptibles: Option[Int] = None
   )(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[ClusterOperationMetadata]] = IO.pure(none[ClusterOperationMetadata])
+  ): IO[Option[OperationSnapshot]] = IO.pure(none[OperationSnapshot])
 
   override def deleteCluster(project: GoogleProject, region: RegionName, clusterName: DataprocClusterName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[ClusterOperationMetadata]] = IO.pure(none[ClusterOperationMetadata])
+  ): IO[Option[OperationSnapshot]] = IO.pure(none[OperationSnapshot])
 
   override def getCluster(project: GoogleProject, region: RegionName, clusterName: DataprocClusterName)(implicit
     ev: Ask[IO, TraceId]