+++ google2/CHANGELOG.md
@@ -5,7 +5,7 @@ This file documents changes to the `workbench-google2` library, including notes 
 ## 0.19
 Changed:
 - Renamed and added fields in `GoogleDataprocService.CreateClusterConfig` to support creating Dataproc clusters with secondary preemptible workers.
-- Changed return type of `GoogleDataprocService.createCluster`
+- Changed return type of `GoogleDataprocService.{createCluster, deleteCluster, resizeCluster}`
 - Removed `RetryConfig` from `GoogleDataprocService` constructors
 
 Added: